# internal/apidocs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sync Backend — Actual API Contract",
    "version": "1.0.0",
    "description": "Source-based contract; registered APIs may be disabled by readiness/config guards. Documentation is read-only. No production activation implied."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host (local or deployed)"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/google": {
      "get": {
        "tags": [
          "Legacy"
        ],
        "summary": "Start browser-cookie OAuth",
        "description": "Browser-only Google OAuth redirect. HttpOnly/Lax state cookie; Secure follows config. offline access; no JSON token response.",
        "operationId": "get__api_v1_auth_google",
        "responses": {
          "302": {
            "description": "Browser redirect. Do not treat as a JSON API.",
            "headers": {
              "Location": {
                "description": "Authorization or mobile completion URL; may contain transient sensitive handoff parameters.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "OAUTH_STATE_INVALID / OAUTH_CODE_MISSING / OAUTH_ACCESS_DENIED / OAUTH_AUTHORIZATION_FAILED: flow rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "OAUTH_STATE_INVALID",
                    "message": "OAuth state is invalid or expired"
                  }
                }
              }
            }
          },
          "401": {
            "description": "OAUTH_SESSION_NOT_FOUND: absent session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "OAUTH_SESSION_NOT_FOUND",
                    "message": "OAuth session was not found"
                  }
                }
              }
            }
          },
          "500": {
            "description": "OAUTH_CONFIGURATION_ERROR / INTERNAL_ERROR: configuration/failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "OAUTH_CONFIGURATION_ERROR",
                    "message": "Google OAuth is not configured"
                  }
                }
              }
            }
          },
          "502": {
            "description": "OAUTH_TOKEN_EXCHANGE_FAILED / YOUTUBE_AUTH_FAILED: provider failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "OAUTH_TOKEN_EXCHANGE_FAILED",
                    "message": "failed to connect Google account"
                  }
                }
              }
            }
          }
        },
        "security": []
      },
      "delete": {
        "tags": [
          "Account"
        ],
        "summary": "Disconnect account",
        "description": "Deletes local token/session, revokes related mobile sessions, clears browser cookies; does not implement Google grant revocation.",
        "operationId": "delete__api_v1_auth_google",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthConnection"
                },
                "example": {
                  "connected": false
                }
              }
            }
          },
          "400": {
            "description": "OAUTH_STATE_INVALID / OAUTH_CODE_MISSING / OAUTH_ACCESS_DENIED / OAUTH_AUTHORIZATION_FAILED: flow rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "OAUTH_STATE_INVALID",
                    "message": "OAuth state is invalid or expired"
                  }
                }
              }
            }
          },
          "500": {
            "description": "OAUTH_CONFIGURATION_ERROR / INTERNAL_ERROR: configuration/failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "OAUTH_CONFIGURATION_ERROR",
                    "message": "Google OAuth is not configured"
                  }
                }
              }
            }
          },
          "502": {
            "description": "OAUTH_TOKEN_EXCHANGE_FAILED / YOUTUBE_AUTH_FAILED: provider failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "OAUTH_TOKEN_EXCHANGE_FAILED",
                    "message": "failed to connect Google account"
                  }
                }
              }
            }
          },
          "401": {
            "description": "MOBILE_SESSION_INVALID / OAUTH_SESSION_NOT_FOUND / YOUTUBE_NOT_CONNECTED: invalid/expired session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "MOBILE_SESSION_INVALID",
                    "message": "a mobile bearer session is required"
                  }
                }
              }
            }
          },
          "409": {
            "description": "AUTH_IDENTITY_CONFLICT: explicit bearer and live browser cookie refer to different sessions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "AUTH_IDENTITY_CONFLICT",
                    "message": "browser and mobile sessions conflict"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "SyncBearer": []
          },
          {
            "BrowserSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthChannel": {
        "properties": {
          "channel_id": {
            "nullable": true,
            "type": "string",
            "description": "null when no YouTube channel resource."
          },
          "channel_title": {
            "nullable": true,
            "type": "string",
            "description": "null when no YouTube channel resource."
          }
        },
        "required": [
          "channel_id",
          "channel_title"
        ],
        "type": "object"
      },
      "AuthConnection": {
        "properties": {
          "connected": {
            "type": "boolean"
          },
          "youtube": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthChannel"
              }
            ],
            "nullable": true,
            "type": "object"
          }
        },
        "required": [
          "connected"
        ],
        "type": "object"
      },
...
      "ModelAPIError": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "ModelErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ModelAPIError"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SyncBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Opaque Sync mobile session (not JWT or Google token)",
        "description": "Obtained from mobile/exchange; no actual token example."
      },
      "BrowserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sync_session",
        "description": "HttpOnly browser cookie set by browser OAuth callback; Android should use SyncBearer."
      }
    }
  }
}